+++ app/pb_kits/playbook/pb_layout/docs/_layout_transparent.jsx
@@ -1,0 +1,24 @@
+import React from 'react'
+import { Layout } from '../../'
+
+const LayoutTransparent = () => {
+  return (
+    <div>
+      <Layout
+          collapse="sm"
+          position="left"
+          size="xs"
+          transparent
+      >
+        <Layout.Side>
+          {'Side'}
+        </Layout.Side>
+        <Layout.Body>
+          {'Body'}
+        </Layout.Body>
+      </Layout>
+    </div>
+  )
+}
+
+export default LayoutTransparent