+++ app/pb_kits/playbook/pb_title_detail/_title_detail.jsx
@@ -1,21 +1,54 @@
+/* @flow */
+
 import React from 'react'
-import PropTypes from 'prop-types'
+import classnames from 'classnames'
 
-const propTypes = {
-  className: PropTypes.string,
-  id: PropTypes.string,
+import {
+  Body,
+  Title,
+} from '../'
+
+import {
+  buildCss,
+  buildDataProps,
+} from '../utilities/props'
+
+type TitleDetailProps = {
+  align?: 'left' | 'center' | 'right',
+  className?: String,
+  data?: object,
+  detail: String,
+  id?: String,
+  title: String,
 }
 
-class TitleDetail extends React.Component {
-  render() {
-    return (
-      <div className="pb_title_detail">
-        <span>{'TITLE DETAIL CONTENT'}</span>
-      </div>
-    )
-  }
-}
+const TitleDetail = ({
+  align = 'left',
+  className,
+  data = {},
+  detail,
+  id,
+  title,
+}: TitleDetailProps) => {
+  const dataProps = buildDataProps(data)
+  const pbCss = buildCss('pb_title_detail_kit', align)
 
-TitleDetail.propTypes = propTypes
+  return (
+    <div
+        {...dataProps}
+        className={classnames(className, pbCss)}
+        id={id}
+    >
+      <Title
+          size={4}
+          text={title}
+      />
+      <Body
+          color="light"
+          text={detail}
+      />
+    </div>
+  )
+}
 
 export default TitleDetail