+++ CHANGELOG.md
@@ -5,8 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [4.15.0] 2020-5-22
 
+### Changed
+- Update filter kit (React) ([#796](https://github.com/powerhome/playbook/pull/796) @thestephenmarshall)
+- Updated React version of Layout Kit to use sub-components ([#801](https://github.com/powerhome/playbook/pull/801)@christinaatai)
+
+### Fixed
+- Sync Ruby and React default 'align' values in DateYearStacked ([#794](https://github.com/powerhome/playbook/pull/794) @kellyeryan)
+- Fix stat change to take numbers outside of 0-100 ([#795](https://github.com/powerhome/playbook/pull/795) @kellyeryan)
+- Don't break button when changing loading from true to false ([#800](https://github.com/powerhome/playbook/pull/800)@gmfvpereira)
+
+### Added
+- Added Progress Step Kit([#776](https://github.com/powerhome/playbook/pull/776) @christinaatai)
+- Created TitleDetail React Playbook Kit ([#783](https://github.com/powerhome/playbook/pull/783) @kellyeryan)
+- Created Timestamp React Playbook Kit ([#786](https://github.com/powerhome/playbook/pull/786) @kellyeryan)
+- Created Online Status React Playbook Kit ([#788](https://github.com/powerhome/playbook/pull/788) @kellyeryan)
+- Added Work-Cell type to contact kit ([#799](https://github.com/powerhome/playbook/pull/799) @kre8sions)
 
 ## [4.14.0] 2020-5-14
 - Upgrade react-popper lib to latest ([#778](https://github.com/powerhome/playbook/pull/778) @thestephenmarshall)